initialised mysql set-up using sequelize
