+++ package.json
@@ -27,8 +27,10 @@
     "express-validator": "^6.11.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.12.12",
+    "mysql2": "^2.2.5",
     "node-redshift": "^0.1.5",
     "nodemon": "^2.0.7",
+    "sequelize": "^6.6.2",
     "swagger-ui-express": "^4.1.6"
   }
 }
