+++ src/db/redshift-conn.js
@@ -1,0 +1,14 @@
+const connectRedshiftDB = () => {
+  var Redshift = require("node-redshift");
+  var options = {};
+  var client = {
+    user: process.env.DB_USER,
+    database: process.env.DB_NAME,
+    password: process.env.DB_PASSWORD,
+    port: process.env.DB_PORT,
+    host: process.env.DB_HOST,
+  };
+  return new Redshift(client, [options]);
+};
+
+module.exports = {connectRedshiftDB };
